+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 2.9.1
+
+### `@liveblocks/client`
+
+- Fix type definition of `ThreadData`: `updatedAt` is always set
+- Fix bug where client wasn't always using the newest delta update backend
+  endpoint yet
+- Fix regression with metadata filtering on explicitly-`undefined` values
+
+### `@liveblocks/react-ui`
+
+- When `Composer` is disabled, its actions are now also disabled as expected.
+- Various event propagation improvements in `Composer`.
+
 ## 2.9.0
 
 We are introducing pagination support to allow apps using threads and inbox